pyproject.toml
[tool.poetry]
name = "sadie-antibody"
version = "v1.0.0" # run 'sadie --version' to get the current version
license = "MIT"
description = "\"The Complete Antibody Library\""
authors = [
    "Jordan R. Willis <jwillis0720@gmail.com>",
    "Troy Sincomb <troysincomb@gmail.com>"
]
readme = "README.md"
homepage = "https://sadie.jordanrwillis.com"
repository = "https://github.com/jwillis0720/sadie"
keywords = ["antibody", "bioinformatics", "biology", "computational biology", "protein", "immunoinformatics"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
]
packages = [
    { include = "sadie", from = "src" },
]

[tool.poetry.urls]
issues = "https://github.com/jwillis0720/sadie/issues"

[tool.poetry.scripts]
sadie = "sadie.app:sadie"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
biopython = "1.80"
click = ">=8.0"
filetype = "^1.1.0"
pandas = ">=1.0,<1.5.0"
requests = "^2.28.1"
pydantic = ">=1.9.0,<2.0.0"
Levenshtein = "^0.20.5"
pyarrow = ">=6.0.1"
semantic-version = "^2.10.0"
PyYAML = "^6.0"
pyhmmer = [
    { version = "^0.7.4", markers= "platform_machine != 'arm64'"  },
    # { git = "https://github.com/althonos/pyhmmer.git", branch = "neon-support", markers= "platform_machine == 'arm64'" }
]
pyhmmer-arm = [{ version = "0.7.5", markers= "platform_machine == 'arm64'"  }]
scikit-learn = "^1.1.2"
openpyxl = "^3.0.10"
yarl = "^1.8.1"
ipython = "^8.5.0"
numpy = "^1.24.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.1.0"
flake8 = "4.0.1"
pre-commit = "2.16.0"
coverage = "6.2"
airr = ">=1.4.0"
black = "22.3.0"
pyinstrument = "^4.3.0"
types-PyYAML = "^6.0.12"
pandas-stubs = "^1.5.0.221003"
types-requests = "^2.28.11.1"
mkdocs = "1.2.3"
mkdocs-material = "8.1.3"
mkdocs-material-extensions = "1.0.3"
mkdocs-git-revision-date-plugin = "0.3.1"
mkdocs-redirects = "1.0.3"
markdown-include = "0.6.0"
mdx-include = "1.4.1"
Jinja2 = "3.0.3"
mypy = "0.982"
bump2version = "1.0.1"
Markdown = "3.3.7"
pyright = ">=1.1.275"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ------------------------------------------------ External Tools ------------------------------------------------ #
[[tool.poetry_bumpversion.replacements]]
files = ["src/sadie/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  | tests/
  | coverage/
  | profiling/
  | src/Bio/
)/
'''

[tool.flake8]
defaultsignore = "B008"

[tool."pylint.MASTER"]
extension-pkg-whitelist = "pydantic"

[tool.coverage.html]
directory = "coverage/reports/html/"

[tool.coverage.xml]
output = "coverage/reports/xml/coverage.xml"

[tool.pyright]
include = ["src"]
exclude = [
    "tests",
    "coverage",
    "src/Bio",
]
reportMissingTypeStubs = false
reportUnkownVariableType = false
reportOptionalIterable = false
# private imports should be allowed. ignore it
reportPrivateImportUsage = false
# TODO: mostly pandas and unknown lists being used
reportGeneralTypeIssues = false
# TODO: list of Nones being used -- should fix this logic
reportOptionalSubscript = false
# TODO: operations being done on list of Nones -- should fix this logic
reportOptionalOperand = false
# TODO: old anacri code using short one time variables
reportUnboundVariable = false
reportMissingImports = false # TODO: should move biopythn 1.79 dep into own folder

[tool.isort]
profile = "black"
known_third_party = ["Bio"]
